requirements.txt
# flask
# gunicorn
# flask-cors
# torch
# opencv-python
# numpy
# matplotlib
# pillow
# PyYAML
# psutil
# scipy
# tqdm
# pandas
# requests

flask
gunicorn
torch
numpy
Pillow